SPI: MIPS: BCM63XX: Simplify TX/RX logic
The hardware TX and RX buffers are "almost" the same size, so the
benefits of using PREPENDBYTE or RO/WO transactions do not justify the
increased driver complexity.  Change the code so that it always uses
full duplex transfers.

This also eliminates the prev/next logic and allows longer chains of
spi_transfer structs, provided that they use the same parameters and
do not exceed the HW buffer sizes.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>