MIPS: Add multiplatform BMIPS target
This supports SMP Linux running on the following SoCs:

 - BCM3384 (BMIPS5000) cable modem application processor, BE
 - BCM3384 (BMIPS4355) cable modem "spare CPU" (experimental), BE
 - BCM6328 (BMIPS4355) ADSL chip, BE
 - BCM7346 (BMIPS5000) set-top box chip, BE
 - BCM7425 (BMIPS5000) set-top box chip, LE

Any big-endian system on the list can boot a common BE image.  Any
little-endian system on the list can boot a common LE image.  On these
MIPS systems, endianness cannot be reconfigured at runtime.

Device Tree is used to configure the following items:

 - UART, USB, GENET peripherals
 - IRQ controllers
 - Early console base address (bcm63xx_uart only)
 - SMP or UP mode
 - MIPS counter frequency
 - Memory size / regions
 - DMA remappings
 - Kernel command line

The DT-enabled bootloader and build instructions for 3384 are posted at
https://github.com/Broadcom/aeolus .  The other chips use legacy non-DT
bootloaders, and so the kernel employs autodetection heuristics to select
a builtin DTB.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>